backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt<4.1
python-dotenv>=1.0.0
pydantic[email]>=2.0.0
python-multipart>=0.0.9
